modified Taskmon,docker files
